Rework conversions.
Allow a filter to replace itself with another filter. This allows
YuvFilter and RGBFilter to automatically skip processing if possible.
And might also be useful in other filters.

Also replaced the ScaleTo method by a SetSize method which returns
'true' if it succeeded and 'false' if it did not. For the sourceFilter
in the CreateFilter method this will always succeed. Also added an
option to ResizeFilter to use SetSize if possible, and rescale only
if this fails. Since this can affect the inputFilter it should be
used carefully.