Replace CombinedChain
Overloading the '+' operator to combine RenderChains gives the same
functionality without needing the extra classes.

This changes makes it possible to add RenderChains (concatenating them)
and to add a RenderChain to a Filter (which will apply the chain to the
filter). For convenience there are also implicit conversions from
Func<IResizeableFilter, IFilter> to RenderChain, and RenderChain to
Func<IFilter, IFilter>.

This commit deprecates: FilterChain, CombinedChain, IRenderChain
(interfaces can't have operators).