Fix rare (?) bug with SuperRes
Using the LumaUpscaler for intermediate scaling caused issues when
anti-ringing was enabled, fixed by using a different scaler.

Also made it compile the shaders up front whether it needs them or not,
I prefer this to suddenly having to compile shaders when the video is
resized. This might cause a delay on startup or when a video starts up,
I'm not too sure which.