Chroma Scaling: Support MPDN's internal chroma scaling. Make ChromaFilter Immutable.
Immutable filters tend to lead to less bugs, although they're slightly
more difficult to deal with. This also makes it easier to use the output
of different chroma scaler inside of a different chroma scaler, which is
convenient for SuperChromaRes.

MPDN's internal chroma scalers are used if no changes are made, but this
functionality is somewhat limited. And performance wise it doesn't seem
to make much of a difference.