Status system overhaul
Keep track of processes by adding tags to the filters.

Tags ignore optimizations, but instead keep track of which filters are used to create a filter.
This avoids problems where a tag could be ignored if the corresponding filter was skipped (as a result of using "ConvertToRgb" or similar).
This behaviour can also be influenced manually (although this can create loops, which cause crashes, so be careful).

In some rare cases ignoring optimizations could show a script is active, even if it isn't technically having any effect.
Usually those cases can be avoided, and are preferrable to the alternative (not displaying something which *is* having an effect).

By default an output is only tagged once (to avoid having 'meta' renderchains show up in the status).
This also makes it possible to override default behaviour.
Care should be taken if the final output of a renderchain is generated by an other renderchain (which never happens except for meta renderchains at the moment).

Tags automatically keep track of subprocesses of a renderchain (things like NNEDI3's chroma scaler, or SuperRes's prescaler).
To make it possible to distinguish between scripts applies sequentially and parallel a 'bottom' tag was added.
This tag which represents the original source of data, but isn't displayed in the status.

Tags remember the entire data flow so it's possible to keep track which input is responsible for which output.
A particular example of this is the special ChromaScalerTag which automatically separates processing done to scale the luma from processing done to scale the chroma and displays the status accordingly.

Tags should continue working no matter how complicated the chain becomes, but the resulting text could be hard to read in more complicated cases.
However unlike the previous system it's reasonably easy to modify the way the tag system is displayed.