[Scripting] Add scaling factor properties
Add "ScalingFactor" to the clip properties, also add "ScalingFactorX"
and "ScalingFactorY" for separate horizontal and vertical scaling
factors.

To make it easier to add automatic properties the IClip interface has
been made into a class. If necessary an interface could be added, but
there seems little need to do so.

Also fixed a bug where the Conditional config dialog tried to select the
wrong renderscript if there are hidden renderscripts.