[WIP] Make AudioChain use Filters
The best way I could find to make this work was to convert "AudioChain"
into a filter and wrap a renderChain around that.

One difficulty is that this makes it harder to handle settings, although
on the flip side you can now assume that settings remain constant

Haven't managed how to connect the "Process" function to the "Render()"
function though, I'm not entirely sure what information is supposed to
go where.

Apart from that the Exemaple scripts are almost done, and the
DynamicRangeCompressor should be more or less finished as well (I hope).