[WIP] Fix hotkey handling
Presets need a better way to let MPDN know that the renderscripts should
be reloaded. Something similar might be needed for the audioscripts.
Either MPDN should do this when "Player.Config.Refresh()" is called, or
perhaps it can be taken care of by adding event handlers.

Currently the DynamicHotkeys class works by registering the 'Hotkeys' as
'Verbs' and letting them be handled by the PlayerExtension class, but it
might make more sense to do this the other way round (which would
simultaneously create a central register of all hotkeys).