[Framework] Create TextureFilter Namespace
Moved Texturefilters to their own namespace.

All dependence of the Renderchain namespace is moved to the file
"RenderChain/TextureFilter/Interface.cs". This makes it easier to keep
track of the dependencies and cleans up the syntax (no more 'new
SomethingFilter'). Downside is that any new functionality will need to
be added to the Interface, on the other hand it's easier to change the
implementation without affecting the RenderChains.

For low level stuff you'll need to use the TextureFilter namespace. But
high-level operations can be done without them. For now OpenCL and
DirectCompute are considered 'low-level'.

Also reworked the ChromaFilter, it's now called 'CompositionFilter' and
must be rebuild every time you make a change to it (somewhat less
flexible than the previous approach but less prone to mistakes).

Also made reworked the way the ShaderPath is determined. It now uses the
project path when debugging (so you can change the files in the project
folder and see the result instantly). And allow external shaders.