[Framework] Change ChromaFilter
Only create chroma scaler on compile (so 'CreateChromaFilter' is only
called once).

This causes some problems with the Tag system, but to solve those we're
going to need a method that can handle the fact that the filter tree can
be rebuild arbitrarily, until it's actually compiled.

It might be *possible* with the current system but it takes more effort
than it should.