[Framework] Revise Tag system
More robust, and can output descriptions of different verbosities (not
used yet, but could e.g. show which shaders are used).

Unforuntately the behaviour of the ChromaScaler tag could not be
replicated (yet). Then again this was frequently a source of bugs, so
it's perhaps for the best.

New output should also be readable with multiple layers of nesting.