[Framework] Chroma Scaling: Rework chroma filter, allow acces to YUV and Luma planes.
Chroma scaling now irrevocably combines Luma and Chroma planes. Meaning
that in a chain of chroma scalers the first is used, rather than the
last, as with the previous methods. This turned out to be a lot simpler.

All chroma scalers now take a CompositionFilter, instead of several
parameters. This CompositionFilter contains the raw YUV information.