[Framework] Fix AudioProcess Loading
Don't allocate GPU resources if process is created before GPU is
initialised (i.e. if we're just loading settings).

Also disabled CudafyInitializer until I can figure out why it's necessary
and why disabling it seems to have no effect (if it's necessary for CS-script
compatibility it should at least be disabled when CS-script isn't used, or maybe
we should abandon CS-script compatibility for MPDN extensions)

Also make AudioProc initialisation asynchronous, and start it as soon as
possible, as it can take quite a bit of time.