[Framework] Fix managed textures, refactor IFilterBase
FilterBase war disposing textures it shouldn't, by creating a
IFilterBase with the texture as output (which automatically disposes the
texture).

Resolved by making "Just" not dispose it's value by default, and
extending "Bound" to allow arbitrary resources to be bound to a
filterBase.  Added new extension method "Bind(x)" which creates a
filterBase with output x, and binds it to x.

This way around is probably clearer than always disposing everything
(which didn't quite match the way IFilterBase was being used).